refactore folder structure
